[HttpClient] Don't pass float to usleep()

+++ src/Symfony/Component/HttpClient/Response/CurlResponse.php
@@ -353,7 +353,7 @@ final class CurlResponse implements ResponseInterface, StreamableInterface
         }
 
         if ($multi->pauseExpiries && 0 < $timeout -= microtime(true) - $now) {
-            usleep(1E6 * $timeout);
+            usleep((int) (1E6 * $timeout));
         }
 
         return 0;
