+++ src/Symfony/Component/HttpClient/Response/NativeResponse.php
@@ -366,7 +366,7 @@ final class NativeResponse implements ResponseInterface, StreamableInterface
         }
 
         if (!$handles) {
-            usleep(1E6 * $timeout);
+            usleep((int) (1E6 * $timeout));
 
             return 0;
         }
